pyproject.toml
[tool.poetry]
name = "curso-fast-api"
version = "0.1.0"
description = ""
authors = ["Guilherme_maciel <aprendiz.guilherme@fontespromotora.com>"]
readme = "README.md"
packages = [{include = "curso_fast_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.91.0"
psycopg2-binary = "^2.9.5"
asyncpg = "^0.27.0"
uvicorn = "^0.20.0"
pydantic = {extras = ["email"], version = "^1.10.5"}
sqlalchemy = "^2.0.3"
sqlalchemy-utils = "^0.40.0"
python-dotenv = "^0.21.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.5"
pytz = "^2022.7.1"
fastapi-limiter = "^0.1.5"
pyrebase4 = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
